+++ front/src/layout/Header.module.scss
@@ -227,17 +227,7 @@
   }
 }
 
-// .headerSignup {
-//   &:hover {
-//     background-color: rgba(0, 0, 0, 0.05) !important;
-//   }
-// }
-
 .headerConnexion {
-  // &:hover {
-  //   background-color: #18009f !important;
-  // }
-
   & span {
     color: #fff;
   }
