+++ front/src/login/Invite.tsx
@@ -1,28 +1,29 @@
 import React, { useState } from "react";
-import { Formik, Form, Field, ErrorMessage } from "formik";
-import * as yup from "yup";
-import { useLocation, Link } from "react-router-dom";
+import { useLocation, useHistory } from "react-router-dom";
 import { useMutation, useQuery, gql } from "@apollo/client";
+import { Formik, Form, Field } from "formik";
+import * as yup from "yup";
 import {
-  Invitation,
   Mutation,
   MutationJoinWithInviteArgs,
   Query,
-  User,
 } from "../generated/graphql/types";
-import {
-  IconEmailActionUnread,
-  IconLock1,
-  IconView,
-  IconSingleNeutralIdCard4,
-} from "common/components/Icons";
 import Loader from "common/components/Loaders";
-import { NotificationError } from "common/components/Error";
-import routes from "common/routes";
-import PasswordMeter from "common/components/PasswordMeter";
-import RedErrorMessage from "common/components/RedErrorMessage";
 import * as queryString from "query-string";
 import { decodeHash } from "common/helper";
+import routes from "common/routes";
+
+import {
+  Container,
+  Row,
+  Col,
+  Title,
+  Text,
+  Alert,
+  Button,
+  TextInput,
+} from "@dataesr/react-dsfr";
+import styles from "./Login.module.scss";
 
 const INVITATION = gql`
   query Invitation($hash: String!) {
@@ -51,75 +52,11 @@ const JOIN_WITH_INVITE = gql`
   }
 `;
 
-/**
- * This page is shown on successful signup
- * The user get a list of all the companies he is part of
- */
-function SignupConfirmation({ user }: { user: User }) {
-  return (
-    <div className="container-narrow">
-      <section className="section section-white">
-        <h2 className="h2 tw-my-4">Confirmation de création de compte</h2>
-        <p className="body-text">
-          Votre compte <span className="tw-font-bold">{user.email}</span> a bien
-          été crée et vous êtes désormais membre des établissements suivants:
-        </p>
-        <ul className="bullets">
-          {user.companies?.map(company => (
-            <li key={company.siret}>
-              {company.name} - ({company.siret})
-            </li>
-          ))}
-        </ul>
-        <p className="body-text">
-          Connectez-vous à votre compte pour accéder à votre tableau de bord et
-          accéder aux bordereaux de ces établissements.
-        </p>
-        <div className="form__actions">
-          <Link to={routes.login} className="btn btn--primary">
-            Se connecter
-          </Link>
-        </div>
-      </section>
-    </div>
-  );
-}
-
-/**
- * This page is shown when user try to join several times with the
- * same invite link
- */
-function AlreadyAccepted({ invitation }: { invitation: Invitation }) {
-  const { email, companySiret } = invitation;
-  return (
-    <div className="container-narrow">
-      <section className="section section-white">
-        <h2 className="h2 tw-my-4">Cette invitation n'est plus valide</h2>
-        <p className="body-text">
-          Votre compte <span className="tw-font-bold">{email}</span> a déjà été
-          crée et le rattachement à l'établissement dont le SIRET est{" "}
-          <span className="tw-font-bold">{companySiret}</span> est effectif.
-        </p>
-        <p className="body-text">
-          Connectez-vous à votre compte pour accéder à votre tableau de bord et
-          accéder aux bordereaux de cet établissement.
-        </p>
-        <div className="form__actions">
-          <Link to={routes.login} className="btn btn--primary">
-            Se connecter
-          </Link>
-        </div>
-      </section>
-    </div>
-  );
-}
-
-/**
- * Signup to Trackdéchets with an invitation link
- */
 export default function Invite() {
+  const [showPassword, setShowPassword] = useState(false);
   // Extract invitation hash from URL
   const location = useLocation();
+  const history = useHistory();
 
   // parse qs and get rid of extra parameters
   const parsedQs = queryString.parse(location.search);
@@ -127,8 +64,6 @@ export default function Invite() {
   const { hash: qsHash } = parsedQs;
 
   const hash = decodeHash(qsHash);
-
-  const [passwordType, setPasswordType] = useState("password");
 
   // INVITATION QUERY
   const { loading, error: queryError, data: queryData } = useQuery<
@@ -149,31 +84,125 @@ export default function Invite() {
     return <Loader />;
   }
 
+  const pageContent = content => {
+    return (
+      <div className={styles.onboardingWrapper}>
+        <Container className={styles.centralContainer} spacing="pt-10w">
+          {content}
+        </Container>
+      </div>
+    );
+  };
+
   if (mutationError) {
-    return <NotificationError apolloError={mutationError} />;
+    return pageContent(
+      <Row justifyContent="center" spacing="mb-2w">
+        <Alert
+          title="Erreur"
+          description={mutationError.message}
+          type="error"
+        />
+      </Row>
+    );
   }
 
   if (mutationData) {
     const user = mutationData.joinWithInvite;
-    return <SignupConfirmation user={user} />;
+
+    return pageContent(
+      <>
+        <Row justifyContent="center" spacing="mb-2w">
+          <Col spacing="m-auto">
+            <Title as="h2" look="h3" spacing="mb-1w">
+              Confirmation de création de compte
+            </Title>
+            <Text as="p" spacing="mb-1w">
+              Votre compte <strong>{user.email}</strong> a bien été crée et vous
+              êtes désormais membre des établissements suivants:
+            </Text>
+            <ul className="bullets">
+              {user.companies?.map(company => (
+                <li key={company.siret}>
+                  {company.name} - ({company.siret})
+                </li>
+              ))}
+            </ul>
+            <Text as="p" spacing="mb-1w">
+              Connectez-vous à votre compte pour accéder à votre tableau de bord
+              et accéder aux bordereaux de ces établissements.
+            </Text>
+          </Col>
+        </Row>
+        <Row justifyContent="right">
+          <Col className={styles.resetFlexCol}>
+            <Button
+              size="md"
+              onClick={() => {
+                history.push({
+                  pathname: routes.login,
+                });
+              }}
+            >
+              Se connecter
+            </Button>
+          </Col>
+        </Row>
+      </>
+    );
   }
 
   if (queryError) {
-    return <NotificationError apolloError={queryError} />;
+    return pageContent(
+      <Row justifyContent="center" spacing="mb-2w">
+        <Alert title="Erreur" description={queryError.message} type="error" />
+      </Row>
+    );
   }
 
   if (queryData && queryData.invitation) {
-    const invitation = queryData.invitation;
+    const { email, companySiret, acceptedAt } = queryData.invitation;
 
-    // invitation was already accepted
-    if (invitation.acceptedAt) {
-      return <AlreadyAccepted invitation={invitation} />;
+    if (acceptedAt) {
+      return pageContent(
+        <>
+          <Row justifyContent="center" spacing="mb-2w">
+            <Col spacing="m-auto">
+              <Title as="h2" look="h3" spacing="mb-1w">
+                Cette invitation n'est plus valide
+              </Title>
+              <Text as="p" spacing="mb-1w">
+                Votre compte <strong>{email}</strong> a déjà été crée et le
+                rattachement à l'établissement dont le SIRET est{" "}
+                <strong>{companySiret}</strong> est effectif.
+              </Text>
+              <Text as="p" spacing="mb-1w">
+                Connectez-vous à votre compte pour accéder à votre tableau de
+                bord et accéder aux bordereaux de ces établissements.
+              </Text>
+            </Col>
+          </Row>
+          <Row justifyContent="right">
+            <Col className={styles.resetFlexCol}>
+              <Button
+                size="md"
+                onClick={() => {
+                  history.push({
+                    pathname: routes.login,
+                  });
+                }}
+              >
+                Se connecter
+              </Button>
+            </Col>
+          </Row>
+        </>
+      );
     }
 
-    return (
+    return pageContent(
       <Formik
         initialValues={{
-          email: invitation?.email ?? "",
+          email: email ?? "",
           name: "",
           password: "",
         }}
@@ -192,99 +221,96 @@ export default function Invite() {
           }).then(_ => setSubmitting(false));
         }}
       >
-        {({ isSubmitting }) => (
-          <section className="section section-white">
-            <div className="container-narrow">
-              <Form>
-                <h1 className="h1 tw-my-4">Validez votre inscription</h1>
-                <p className="body-text">
+        {({ isSubmitting, errors, touched, isValid, submitForm }) => (
+          <Form>
+            <Row justifyContent="center" spacing="mb-2w">
+              <Col spacing="m-auto">
+                <Title as="h1" look="h3" spacing="mb-1w">
+                  Validez votre inscription
+                </Title>
+                <Text as="p" spacing="mb-1w">
                   Vous avez été invité à rejoindre Trackdéchets. Pour valider
                   votre inscription, veuillez compléter le formulaire
                   ci-dessous.
-                </p>
-
-                <div className="form__row">
-                  <label>Email</label>
-                  <div className="field-with-icon-wrapper">
-                    <Field
-                      type="email"
-                      name="email"
-                      className="td-input"
-                      readOnly
-                    />
-                    <i>
-                      <IconEmailActionUnread />
-                    </i>
-                  </div>
-                  <RedErrorMessage name="email" />
-                </div>
-
-                <div className="form__row">
-                  <label>Nom et prénom</label>
-                  <div className="field-with-icon-wrapper">
-                    <Field type="text" name="name" className="td-input" />
-                    <i>
-                      <IconSingleNeutralIdCard4 />
-                    </i>
-                  </div>
-
-                  <RedErrorMessage name="name" />
-                </div>
-
-                <div className="form__row">
-                  <label>Mot de passe</label>
-
-                  <Field name="password">
-                    {({ field }) => {
-                      return (
-                        <>
-                          <div className="field-with-icon-wrapper">
-                            <input
-                              type={passwordType}
-                              {...field}
-                              className="td-input"
-                            />
-                            <i>
-                              <IconLock1 />
-                            </i>
-                          </div>
-                          <span
-                            className="showPassword"
-                            onClick={() =>
-                              setPasswordType(
-                                passwordType === "password"
-                                  ? "text"
-                                  : "password"
-                              )
-                            }
-                          >
-                            <IconView /> <span>Afficher le mot de passe</span>
-                          </span>
-                          <PasswordMeter password={field.value} />
-                        </>
-                      );
-                    }}
-                  </Field>
-
-                  <RedErrorMessage name="password" />
-                </div>
-
-                <ErrorMessage
-                  name="passwordConfirmation"
-                  render={msg => <div className="error-message">{msg}</div>}
-                />
-                <div className="form__actions">
-                  <button
-                    className="btn btn--primary"
-                    type="submit"
-                    disabled={isSubmitting}
-                  >
-                    Valider l'inscription
-                  </button>
-                </div>
-              </Form>
-            </div>
-          </section>
+                </Text>
+                <Text as="p" className="fr-text--bold">
+                  Vos informations :
+                </Text>
+                <Field name="name">
+                  {({ field }) => {
+                    return (
+                      <TextInput
+                        // @ts-ign
+                        {...field}
+                        required
+                        label="Nom et prénom"
+                        messageType={errors.name && touched.name ? "error" : ""}
+                        message={errors.name && touched.name ? errors.name : ""}
+                      />
+                    );
+                  }}
+                </Field>
+                <Field name="email">
+                  {({ field }) => {
+                    return (
+                      <TextInput
+                        // @ts-ignore
+                        {...field}
+                        readOnly
+                        required
+                        label="Email"
+                      />
+                    );
+                  }}
+                </Field>
+                <Field name="password">
+                  {({ field }) => {
+                    return (
+                      <TextInput
+                        type={showPassword ? "text" : "password"}
+                        {...field}
+                        required
+                        label="Mot de passe"
+                        messageType={
+                          errors.password && touched.password ? "error" : ""
+                        }
+                        message={
+                          errors.password && touched.password
+                            ? errors.password
+                            : ""
+                        }
+                      />
+                    );
+                  }}
+                </Field>
+                <Button
+                  tertiary
+                  hasBorder={false}
+                  icon={showPassword ? "ri-eye-off-line" : "ri-eye-line"}
+                  iconPosition="left"
+                  onClick={() => setShowPassword(!showPassword)}
+                >
+                  Afficher le mot de passe
+                </Button>
+              </Col>
+            </Row>
+            <Row justifyContent="right">
+              <Col className={styles.resetFlexCol}>
+                <Button
+                  icon="ri-arrow-right-line"
+                  iconPosition="right"
+                  size="md"
+                  onClick={submitForm}
+                  disabled={!isValid}
+                  title={
+                    isSubmitting ? "Création en cours..." : "Créer mon compte"
+                  }
+                >
+                  Créer mon compte
+                </Button>
+              </Col>
+            </Row>
+          </Form>
         )}
       </Formik>
     );
