+++ front/src/login/Login.module.scss
@@ -6,8 +6,8 @@
       max-width: 688px;
     }
   }
-}
 
-.resetFlexCol {
-  flex: 0 1 auto;
+  .resetFlexCol {
+    flex: 0 1 auto;
+  }
 }
