+++ front/src/login/Signup.tsx
@@ -1,207 +1,228 @@
-import { useMutation } from "@apollo/client";
-import { Field } from "formik";
-import React, { useState } from "react";
+import React, { useState, createRef } from "react";
 import { useHistory } from "react-router-dom";
-import {
-  IconEmailActionUnread,
-  IconView,
-  IconSingleNeutralIdCard4,
-  IconLock1,
-  IconPhone,
-} from "common/components/Icons";
-import { NotificationError } from "common/components/Error";
-import PasswordMeter from "common/components/PasswordMeter";
-import RedErrorMessage from "common/components/RedErrorMessage";
-import { SIGNUP } from "./mutations";
-import { Wizard } from "./Wizard";
+import { useMutation } from "@apollo/client";
 import { Mutation, MutationSignupArgs } from "generated/graphql/types";
+import { SIGNUP } from "./mutations";
 import routes from "common/routes";
 
+import {
+  Container,
+  Row,
+  Col,
+  Title,
+  Text,
+  TextInput,
+  Button,
+  Checkbox,
+  Link,
+  Alert,
+} from "@dataesr/react-dsfr";
+import styles from "./Login.module.scss";
+
+import { CONTACT_EMAIL } from "common/config";
+
 export default function Signup() {
-  const [passwordType, setPasswordType] = useState("password");
-  const [signup, { error: signupError }] = useMutation<
-    Pick<Mutation, "signup">,
-    MutationSignupArgs
-  >(SIGNUP);
+  const [showPassword, setShowPassword] = useState(false);
+  const [submittable, setSubmittable] = useState(false);
+  const [submitting, setSubmitting] = useState(false);
+  const [signupCompleted, setSignupCompleted] = useState(false);
+  const [errorMessage, setErrorMessage] = useState("");
+  const [userEmail, setUserEmail] = useState("");
+
+  const [signup] = useMutation<Pick<Mutation, "signup">, MutationSignupArgs>(
+    SIGNUP
+  );
 
   const history = useHistory();
 
-  return (
-    <section className="section section--white">
-      <div className="container">
-        <Wizard
-          initialValues={{
-            email: "",
-            name: "",
-            phone: "",
-            password: "",
-            cgu: false,
-          }}
-          onSubmit={(values: any, { setSubmitting }) => {
-            const { cgu, ...userInfos } = values;
+  const nameRef = createRef<HTMLInputElement>();
+  const emailRef = createRef<HTMLInputElement>();
+  const passwordRef = createRef<HTMLInputElement>();
+  const cguRef = createRef<HTMLInputElement>();
 
-            signup({ variables: { userInfos } })
-              .then(_ =>
-                history.push({
-                  pathname: routes.signup.activation,
-                  state: { signupEmail: userInfos.email },
-                })
-              )
-              .catch(_ => {
-                setSubmitting(false);
-              });
-          }}
-        >
-          <Wizard.Page title="Bienvenue" formClassName="container-narrow">
-            <div>
-              <h1 className="h1 tw-mb-6">Bienvenue sur Trackdéchets</h1>
+  const onSubmit = () => {
+    const userInfos = {
+      email: emailRef.current?.value || "",
+      name: nameRef.current?.value || "",
+      password: passwordRef.current?.value || "",
+    };
 
-              <p className="body-text">
-                Trackdéchets est un produit du Ministère de la Transition
-                Ecologique et Solidaire.
-              </p>
+    setSubmitting(true);
 
-              <p className="body-text">
-                Son objectif : simplifier la gestion des déchets dangereux au
-                quotidien. 0 papier, traçabilité en temps réel, informations
-                regroupées sur un outil unique, vérification de vos
-                prestataires.
-              </p>
+    signup({ variables: { userInfos } })
+      .then(_ => {
+        setUserEmail(userInfos.email);
+        setSignupCompleted(true);
+      })
+      .catch(_ => {
+        setErrorMessage(
+          _.message || "Une erreur est survenue, veuillez réessayer."
+        );
+        setSubmitting(false);
+      });
+  };
 
-              <p className="body-text">
-                Il est libre d'utilisation et utilisable par tous les acteurs de
-                la filière déchets. Rejoignez-nous !
-              </p>
+  const onConnectClick = () => {
+    history.push({
+      pathname: routes.login,
+    });
+  };
 
-              <p className="body-text">
-                Vous vous apprêtez à créer un compte personnel. Cette étape est
-                un prélable obligatoire à l'enregistrement ou au rattachement
-                d'une entreprise dans Trackdéchets.
-              </p>
-            </div>
-          </Wizard.Page>
-          <Wizard.Page
-            title="Informations utilisateur"
-            formClassName="container-narrow"
-            validate={(values: any) => {
-              let errors: any = {};
+  const onChange = () => {
+    const formFilled =
+      !!nameRef.current?.value &&
+      !!emailRef.current?.value &&
+      !!passwordRef.current?.value &&
+      !!cguRef.current?.checked;
 
-              if (!values.email) {
-                errors.email = "L'email est obligatoire";
-              }
-              if (!values.name) {
-                errors.name = "Le nom et prénom sont obligatoires";
-              }
+    setSubmittable(formFilled);
+  };
 
-              if (!values.password) {
-                errors.password = "Le mot de passe ne peut pas être vide";
-              }
+  const alert =
+    errorMessage.length > 0 ? (
+      <Row spacing="mb-2w">
+        <Alert title="Erreur" description={errorMessage} type="error" />
+      </Row>
+    ) : null;
 
-              if (!values.cgu) {
-                errors.cgu =
-                  "Vous devez avoir lu les conditions générales d'utilisation";
-              }
-
-              return errors;
-            }}
+  const formContent = (
+    <Container className={styles.centralContainer} spacing="pt-10w">
+      {alert}
+      <Row justifyContent="center" spacing="mb-2w">
+        <Col spacing="m-auto">
+          <Title as="h1" look="h3" spacing="mb-1w">
+            Créer mon compte Trackdéchets
+          </Title>
+          <Text as="p" spacing="mb-1w">
+            Vous vous apprêtez à créer votre compte utilisateur. Cette étape est
+            préalable à l'enregistrement ou au rattachement d'une entreprise
+            dans Trackdéchets.
+          </Text>
+          <Text as="p" className="fr-text--bold">
+            Vos informations :
+          </Text>
+          <TextInput
+            // @ts-ignore Ref isn't part of the interface
+            ref={nameRef}
+            required
+            label="Nom et prénom"
+            onBlur={onChange}
+          />
+          <TextInput
+            // @ts-ignore ref
+            ref={emailRef}
+            required
+            label="Email"
+            onBlur={onChange}
+          />
+          <TextInput
+            type={showPassword ? "text" : "password"}
+            required
+            label="Mot de passe"
+            // @ts-ignore
+            ref={passwordRef}
+            onBlur={onChange}
+          />
+          <Button
+            tertiary
+            hasBorder={false}
+            icon={showPassword ? "ri-eye-off-line" : "ri-eye-line"}
+            iconPosition="left"
+            onClick={() => setShowPassword(!showPassword)}
           >
-            {" "}
-            <div className="container-narrow">
-              <h1 className="h1 tw-mb-6">Informations utilisateur</h1>
+            {showPassword ? "Masquer" : "Afficher"} le mot de passe
+          </Button>
+        </Col>
+      </Row>
+      <Row spacing="mb-2w">
+        <Col className={styles.resetFlexCol}>
+          <Checkbox
+            // @ts-ignore no change event in interface / passes as remaining props
+            onChange={onChange}
+            label="Je certifie avoir lu les conditions générales"
+            ref={cguRef}
+          />
+          <Link
+            href="https://trackdechets.beta.gouv.fr/cgu"
+            target="_blank"
+            isSimple
+            // @ts-ignore
+            rel="noopener noreferrer"
+          >
+            Voir les conditions générales.
+          </Link>
+        </Col>
+      </Row>
+      <Row justifyContent="right">
+        <Col className={styles.resetFlexCol}>
+          <Button
+            icon="ri-arrow-right-line"
+            iconPosition="right"
+            size="md"
+            title={submitting ? "Création en cours..." : "Créer mon compte"}
+            disabled={!submittable || submitting}
+            onClick={onSubmit}
+          >
+            Créer mon compte
+          </Button>
+        </Col>
+      </Row>
+    </Container>
+  );
 
-              <div className="form__row">
-                <label>Nom et prénom</label>
-                <div className="field-with-icon-wrapper">
-                  <Field type="text" name="name" className="td-input" />
-                  <i>
-                    <IconSingleNeutralIdCard4 />
-                  </i>
-                </div>
+  const successContent = (
+    <Container className={styles.centralContainerLarge} spacing="pt-10w">
+      <Row justifyContent="center" spacing="mb-2w">
+        <Col spacing="m-auto pr-2w">
+          <Title as="h1" look="h3" spacing="mb-1w">
+            On y est presque !
+          </Title>
+          <Text as="p" spacing="mb-1w">
+            Un email de confirmation vous a été envoyé à l'adresse{" "}
+            <strong>{userEmail}</strong> 📨
+          </Text>
+          <Text as="p">👉 Il peut mettre quelques minutes à arriver</Text>
+          <Text as="p">👉 Vérifiez vos spams ou indésirables</Text>
+          <Text as="p">👉 Ajouter {CONTACT_EMAIL} à vos contacts</Text>
+          <Text as="p">
+            👉 Si vous n'avez pas reçu l'email de confirmation au bout d'une
+            heure, vous pouvez le renvoyer depuis{" "}
+            <Link href={routes.resendActivationEmail} isSimple>
+              cette page
+            </Link>
+          </Text>
+          <Text as="p">
+            Le message peut ne pas arriver pour les raisons suivantes :<br />-
+            adresse email erronée
+            <br />- antivirus ou suite logicielle de sécurité trop restrictifs
+          </Text>
+          <Text as="p">
+            Pour finaliser votre inscription, cliquez sur le lien qui vous a été
+            envoyé par email. Vous pourrez ensuite vous connecter à
+            Trackdéchets. 🚀
+          </Text>
+          <Text as="p">
+            Des questions, des interrogations ? N'hésitez pas à{" "}
+            <Link href={`mailto:${CONTACT_EMAIL}`} isSimple>
+              nous contacter
+            </Link>
+            .
+          </Text>
+        </Col>
+      </Row>
+      <Row justifyContent="right">
+        <Col className={styles.resetFlexCol}>
+          <Button size="md" title="Se connecter" onClick={onConnectClick}>
+            Se connecter
+          </Button>
+        </Col>
+      </Row>
+    </Container>
+  );
 
-                <RedErrorMessage name="name" />
-              </div>
-
-              <div className="form__row">
-                <label>Email</label>
-                <div className="field-with-icon-wrapper">
-                  <Field type="email" name="email" className="td-input" />
-                  <i>
-                    <IconEmailActionUnread />
-                  </i>
-                </div>
-
-                <RedErrorMessage name="email" />
-              </div>
-
-              <div className="form__row">
-                <label>Téléphone (optionnel)</label>
-                <div className="field-with-icon-wrapper">
-                  <Field type="text" name="phone" className="td-input" />
-                  <i>
-                    <IconPhone />
-                  </i>
-                </div>
-              </div>
-
-              <div className="form__row">
-                <label>Mot de passe</label>
-
-                <Field name="password">
-                  {({ field }) => {
-                    return (
-                      <>
-                        <div className="field-with-icon-wrapper">
-                          <input
-                            type={passwordType}
-                            {...field}
-                            className="td-input"
-                          />
-                          <i>
-                            <IconLock1 />
-                          </i>
-                        </div>
-                        <span
-                          className="showPassword"
-                          onClick={() =>
-                            setPasswordType(
-                              passwordType === "password" ? "text" : "password"
-                            )
-                          }
-                        >
-                          <IconView /> <span>Afficher le mot de passe</span>
-                        </span>
-                        <PasswordMeter password={field.value} />
-                      </>
-                    );
-                  }}
-                </Field>
-
-                <RedErrorMessage name="password" />
-              </div>
-
-              <div className="form__row">
-                <label>
-                  <Field name="cgu" type="checkbox" className="td-checkbox" />
-                  Je certifie avoir lu les{" "}
-                  <a
-                    href="https://trackdechets.beta.gouv.fr/cgu"
-                    target="_blank"
-                    className="link"
-                    rel="noopener noreferrer"
-                  >
-                    conditions générales d'utilisation
-                  </a>
-                  *
-                </label>
-              </div>
-
-              <RedErrorMessage name="cgu" />
-
-              {signupError && <NotificationError apolloError={signupError} />}
-            </div>
-          </Wizard.Page>
-        </Wizard>
-      </div>
-    </section>
+  return (
+    <div className={styles.onboardingWrapper}>
+      {signupCompleted ? successContent : formContent}
+    </div>
   );
 }
