--- front/src/login/SignupInfo.module.scss
@@ -1,4 +1,0 @@
-.signupInfoContainer {
-  max-width: 40rem;
-  margin: 0 auto;
-}
