+++ front/src/scss/index.scss
@@ -1,6 +1,4 @@
-.App {
-  @import "../generated/css/tailwind.generated.css";
-}
+@import "../generated/css/tailwind.generated.css";
 @import "./Constants.scss";
 @import "./Breakpoints.scss";
 @import "./Typography.scss";
